// src/interfaces/result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EU Digital COVID Certificate Validation Decorator
 * The API provides functionalities for validating EU digital COVID certificates.
 *
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Result
 */
export interface Result {
    /**
     * 
     * @type {string}
     * @memberof Result
     */
    identifier?: any;
    /**
     * 
     * @type {string}
     * @memberof Result
     */
    result?: any;
    /**
     * 
     * @type {string}
     * @memberof Result
     */
    type?: any;
    /**
     * 
     * @type {string}
     * @memberof Result
     */
    details?: any;
}